requirements.txt
# Core data science libraries
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.5.0  # Interactive visualizations

# Financial data
yfinance>=0.1.70
alpha_vantage>=2.3.1  # Alpha Vantage API client
polygon-api-client>=1.6.0  # Polygon.io API client (optional)
pandas-datareader>=0.10.0  # Additional data sources

# Real-time data and networking
websocket-client>=1.3.3  # For WebSocket connections
requests>=2.28.0  # For REST API calls
pytz>=2022.1  # For timezone handling

# Backtesting and trading
backtrader>=1.9.76.123  # Optional - alternative backtesting engine
pyfolio>=0.9.2  # Portfolio and risk analytics
ta>=0.10.1  # Technical analysis indicators

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
pandas_market_calendars>=3.2  # Market calendars for trading hours
pyyaml>=6.0  # For YAML configuration files

# Email and notifications
secure-smtplib>=0.1.1  # For email notifications

# Optimization
scipy>=1.7.0  # Scientific computing
hyperopt>=0.2.7  # Hyperparameter optimization

# Optional - for advanced features
# xgboost>=1.5.0
# lightgbm>=3.3.0
# tensorflow>=2.8.0
# statsmodels>=0.13.0  # Statistical models
# arch>=5.0.0  # ARCH models for volatility